requirements.txt
Django==1.6.5
argparse==1.2.1
django-crispy-forms==1.4.0
django-reversion==1.8.1
django-xadmin==0.5.0
wsgiref==0.1.2
